cleaning up the pod utility functions into their own pkg
