+++ pkg/controller/leaderworkerset_controller_test.go
@@ -383,81 +383,6 @@ func TestConstructWorkerStatefulSetApplyConfiguration(t *testing.T) {
 	}
 }
 
-func TestContainerRestared(t *testing.T) {
-	tests := []struct {
-		name                    string
-		pod                     corev1.Pod
-		expectRestaredContainer bool
-	}{
-		{
-			name: "Pod in running phase, InitContainerStatuses has restart count > 0",
-			pod: corev1.Pod{
-				Status: corev1.PodStatus{
-					Phase: corev1.PodRunning,
-					InitContainerStatuses: []corev1.ContainerStatus{{
-						RestartCount: 1,
-					}},
-				},
-			},
-			expectRestaredContainer: true,
-		},
-		{
-			name: "Pod in pending phase, InitContainerStatuses has restart count > 0",
-			pod: corev1.Pod{
-				Status: corev1.PodStatus{
-					Phase: corev1.PodPending,
-					InitContainerStatuses: []corev1.ContainerStatus{{
-						RestartCount: 1,
-					}},
-				},
-			},
-			expectRestaredContainer: true,
-		},
-		{
-			name: "Pod in running phase, ContainerStatuses has restart count > 0",
-			pod: corev1.Pod{
-				Status: corev1.PodStatus{
-					Phase: corev1.PodRunning,
-					ContainerStatuses: []corev1.ContainerStatus{{
-						RestartCount: 1,
-					}},
-				},
-			},
-			expectRestaredContainer: true,
-		},
-		{
-			name: "Pod in Failed status",
-			pod: corev1.Pod{
-				Status: corev1.PodStatus{
-					Phase: corev1.PodFailed,
-				},
-			},
-		},
-		{
-			name: "Pod in running phase, InitContainerStatuses has restart count = 0, ContainerStatuses = 0",
-			pod: corev1.Pod{
-				Status: corev1.PodStatus{
-					Phase: corev1.PodRunning,
-					InitContainerStatuses: []corev1.ContainerStatus{{
-						RestartCount: 0,
-					}},
-					ContainerStatuses: []corev1.ContainerStatus{{
-						RestartCount: 0,
-					}},
-				},
-			},
-		},
-	}
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			containerRestared := containerRestarted(tc.pod)
-			if containerRestared != tc.expectRestaredContainer {
-				t.Errorf("Expected value %t, got %t", tc.expectRestaredContainer, containerRestared)
-			}
-		})
-	}
-}
-
 func TestExclusiveConditionTypes(t *testing.T) {
 	tests := []struct {
 		name                          string
